fix(tui): chat: remove paste key binding from onboarding
The keybinding does not bind to any action in the app but rather assumes
the terminal application's default paste behavior. We shouldn't assume
that and instead leave it to the terminal emulator and user to handle
pasting.

On most macOS terminals the default paste keybinding is `cmd+v` and on
most Windows and Linux terminals it is `ctrl+shift+v`.