fix(onboarding): fix y/n key press on model list and api key field
A bug on onboarding made it impossible to press "y" and "n" on
onboarding, on both the model list and the API key field.

This also means that on some terminals, like the Windows Terminal,
pasting an API key would supress all "y" and "n" chars present in the
key, because these terminal receives an even for each character being
pressed.