sec: remove 'go' from safe programs
It could be used to prompt inject commands e.g. `go test -exec`.

Plus, we don't have other languages there, so I think we remove Go as
well.

Ideally, we could ask haiku if the command is read only or not, and ask
for perms based on that, but for now I think this will do.

Thanks Will Vandevanter for the report.

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>