fix(tui): disable progress bar for iterm2 and over ssh
This will conditionally query the terminal for its version and name on
startup. If the program is run locally or the TERM_PROGRAM variable is
set to a non-Apple terminal, Crush will query for the terminal version.
Otherwise, if the terminal is known to support XTVERSION, it will also
query for the version.

The problem is that Apple Terminal does not support XTVERSION, and when
we query for XTVERSION, it render some jargon that might break the UI.

Once we get a response, we send progress-bar update messages only if the
terminal is _not_ iTerm2 since it doesn't support that yet.

Fixes: https://github.com/charmbracelet/crush/issues/1326