feat: add ability to set envs to override config directories
* `CRUSH_GLOBAL_CONFIG`
* `CRUSH_GLOBAL_DATA`

This is more for us than for the end users. It makes it a lot easier to
test the onboarding flow, setting up a providers, etc. You can just do
something like this:

```bash
env CRUSH_GLOBAL_DATA=tmp/data crush
```

And it'll simulate Crush being opened for the first time on the machine.